cambiado estado a nombre, no a int

+++ desguaceNET/desguaceNET/DesguaceNet.cs
@@ -297,7 +297,7 @@ namespace desguaceNET
                 realizado = bd.cambiarEstadoPedido(estado, id);
                 if (realizado)
                 {
-                    cambiarEstadoPedido(id, ((int)estado).ToString());
+                    cambiarEstadoPedido(id, ((estado).ToString());
                 }
 
                 return realizado;
